2.1.0
Все пункты справлены кроме 1, название qqq

+++ qqq/Dtos/CarDto.cs
@@ -2,8 +2,10 @@
 {
     public class CarDto
     {
-        public int Id { get; set; }
+       
+        public long Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
+
     }
 }
